+++ tg_sdk/client/__init__.py
@@ -1,3 +1,0 @@
-from tg_sdk.client.client import Client
-
-__all__ = ["Client"]