+++ tg_sdk/order/__init__.py
@@ -1,6 +1,0 @@
-from tg_sdk.order.order import Order
-from tg_sdk.order.exceptions import InvalidItemsException
-
-__all__ = ['Order',
-           'InvalidItemsException',
-           'InvalidCustomerInformationException', ]