+++ tg_sdk/policy/__init__.py
@@ -1,8 +1,0 @@
-from tg_sdk.policy.constants import UPGRADED
-from tg_sdk.policy.policy import Policy
-from tg_sdk.policy.exceptions import NoBillingAddressException
-
-__all__ = [
-    "Policy",
-    "NoBillingAddressException",
-    "UPGRADED", ]